+++ RELEASE_NOTES.rst
@@ -36,6 +36,7 @@ Highlights:
   in particular to use transparent compression filters as follows:
   ``set_dataset(name, value, hdf5_options={"compression": "gzip"})``.
 * Removed worker DB warning for writing a dataset that is also in the archive
+* The "ip" config option can now be set to "use_dhcp" to use DHCP to obtain an IP address.
 
 Breaking changes: